# cs309/Documents/backend-javadoc/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api Documentation",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://coms-309-063.cs.iastate.edu:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/trips": {
      "get": {
        "tags": [
          "trip-controller"
        ],
        "summary": "getAllTrips",
        "operationId": "getAllTripsUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TripRes"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "trip-controller"
        ],
        "summary": "createTrip",
        "operationId": "createTripUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{id}": {
      "get": {
        "tags": [
          "trip-controller"
        ],
        "summary": "getTripById",
        "operationId": "getTripByIdUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TripRes"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "trip-controller"
        ],
        "summary": "updateTrip",
        "operationId": "updateTripUsingPUT",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TripRes"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "trip-controller"
        ],
        "summary": "deleteTrip",
        "operationId": "deleteTripUsingDELETE",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/users/assign-trip/{userId}/trip/{tripId}": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "assignTripToUser",
        "operationId": "assignTripToUserUsingPUT",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "tripId",
            "in": "path",
            "description": "tripId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/create": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "createUser",
        "operationId": "createUserUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/users/get": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getAllUsers",
        "operationId": "getAllUsersUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRes"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/users/get/id/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getUserById",
        "operationId": "getUserByIdUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserRes"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/users/{id}": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "deleteTrip",
        "operationId": "deleteTripUsingDELETE_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Comment": {
        "title": "Comment",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "FlightReq": {
        "title": "FlightReq",
        "type": "object",
        "properties": {
          "airlineName": {
            "type": "string"
          },
          "arriving": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "departing": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "trip": {
            "$ref": "#/components/schemas/TripReq"
          }
        }
      },
      "FlightRes": {
        "title": "FlightRes",
        "type": "object",
        "properties": {
          "airlineName": {
            "type": "string"
          },
          "arriving": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "departing": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Lodging": {
        "title": "Lodging",
        "type": "object",
        "properties": {
          "checkIn": {
            "type": "string"
          },
          "checkOut": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "location": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "type": "string",
            "enum": [
              "BNB",
              "CABIN",
              "HOTEL"
            ]
          }
        }
      },
...
      "Post": {
        "title": "Post",
        "type": "object",
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "date": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string"
          },
          "likesCounter": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "TripReq": {
        "title": "TripReq",
        "type": "object",
        "properties": {
          "duration": {
            "type": "string"
          },
          "flight": {
            "$ref": "#/components/schemas/FlightReq"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isArchived": {
            "type": "boolean"
          },
          "location": {
            "type": "string"
          },
          "lodging": {
            "$ref": "#/components/schemas/Lodging"
          }
        }
      },
      "TripRes": {
        "title": "TripRes",
        "type": "object",
        "properties": {
          "duration": {
            "type": "string"
          },
          "flight": {
            "$ref": "#/components/schemas/FlightRes"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isArchived": {
            "type": "boolean"
          },
          "location": {
            "type": "string"
          },
          "lodging": {
            "$ref": "#/components/schemas/Lodging"
          }
        }
      },
      "UserReq": {
        "title": "UserReq",
        "type": "object",
        "properties": {
          "accountType": {
            "type": "string",
            "enum": [
              "ADMIN",
              "BASIC",
              "POSTER"
            ]
          },
          "emailId": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "joiningDate": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "trips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripReq"
            }
          }
        }
      },
      "UserRes": {
        "title": "UserRes",
        "type": "object",
        "properties": {
          "accountType": {
            "type": "string",
            "enum": [
              "ADMIN",
              "BASIC",
              "POSTER"
            ]
          },
          "emailId": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "joiningDate": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }

          },
          "trips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripRes"
            }
          }
        }
      },
...
    }
  }
}