# cs309/Experiments/saiyaraift/ASBasicAPIcalls/app/src/main/java/com/example/basicapicalls/model/Post.java
package com.example.basicapicalls.model;

import com.google.gson.annotations.SerializedName;

public class Post{
    private int userID;
    private int id;
    private String title;
    @SerializedName("body")
    private String bigText;

    public int getUserID() {
        return userID;
    }

    public void setUserID(int userID) {
        this.userID = userID;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getBigText() {
        return bigText;
    }

    public void setBigText(String bigText) {
        this.bigText = bigText;
    }
}